+++ packages/helper/src/index.ts
@@ -1,10 +1,14 @@
-export * from "./types/array";
-export * from "./types/env";
-export * from "./types/json";
-export * from "./types/string";
+import * as array from "./types/array";
+import * as env from "./types/env";
+import * as json from "./types/json";
+import * as string from "./types/string";
+import * as path from "./types/path";
 
-export * from "./types/path";
+import { Optional } from "./types/optional";
 
 import stream from "./types/stream";
 
-export { stream };
+// eslint-disable-next-line @typescript-eslint/naming-convention
+const data = { Optional };
+
+export { env, array, json, string, path, stream, data };