+++ packages/helper/src/test/array.test.ts
@@ -1,61 +1,24 @@
-import { jsonToArray } from "..";
+import { array } from "..";
 
-describe("Array Helper", () => {
-  test("empty json", () => {
-    const json = {};
-    const array = jsonToArray(json);
-
-    expect(array).toHaveLength(0);
+describe("Array helper", () => {
+  test("change string to string array", () => {
+    expect(array.toArray("string")).toHaveLength(1);
   });
 
-  test("partial json not filter empty string", () => {
-    const json = { one: { index: 1, data: "string" }, two: undefined };
-    const array = jsonToArray(json);
+  test("change object to object array", () => {
+    const obj = { a: 12, b: "hello" };
+    const arr = array.toArray(obj);
 
-    expect(array).toHaveLength(2);
-    expect(array).toContain("string");
+    expect(arr).toHaveLength(1);
+    expect(arr[0]).toEqual(obj);
   });
 
-  test("multiple partial json", () => {
-    const json = {
-      one: { index: 1, data: "string" },
-      two: undefined,
-      three: undefined,
-      four: undefined,
-      five: { index: 2, data: "number" },
-    };
-    const array = jsonToArray(json);
+  test("flatmap array", () => {
+    const str = "asdf";
+    const a1: (string | string[])[] = [str, str, str, [str], str, [str, str, str]];
+    const a2: (string | string[])[] = [[str, str], str, str, [str, str]];
+    const arr = array.flatmap(a1, a2);
 
-    expect(array).toHaveLength(5);
-    expect(array).toContain("string");
-    expect(array).toContain("number");
-  });
-
-  test("complete json", () => {
-    const json = {
-      two: { index: 1, data: "string" },
-      one: { index: 3, data: "number" },
-      three: { index: 2, data: "boolean" },
-    };
-    const array = jsonToArray(json);
-
-    expect(array).toHaveLength(3);
-    expect(array[0]).toEqual("string");
-    expect(array[1]).toEqual("boolean");
-    expect(array[2]).toEqual("number");
-  });
-
-  test("complete json with data as string[]", () => {
-    const json = {
-      two: { index: 1, data: ["string", "number", "boolean"] },
-      one: { index: 3, data: ["number3", "string3"] },
-      three: { index: 2, data: "boolean2" },
-    };
-    const array = jsonToArray(json);
-
-    expect(array).toHaveLength(6);
-    expect(array[0]).toEqual("string");
-    expect(array[3]).toEqual("boolean2");
-    expect(array[4]).toEqual("number3");
+    expect(arr).toHaveLength(14);
   });
 });