+++ packages/logger/src/models/output/OutputMessagePrefix.ts
@@ -1,0 +1,7 @@
+import { json } from "@kcutils/helper";
+
+export interface OutputMessagePrefix extends json.OptionalSortableJson {
+  badge: json.JsonSortableData<string>;
+  label: json.JsonSortableData<string>;
+  custom: json.JsonSortableData<string[]>;
+}