+++ packages/logger/src/models/output/OutputMessageSuffix.ts
@@ -1,0 +1,5 @@
+import { json } from "@kcutils/helper";
+
+export interface OutputMessageSuffix extends json.OptionalSortableJson {
+  custom: json.JsonSortableData<string[]>;
+}