+++ internals/react-testkit/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3-rc.1](https://github.com/kamontat/kcutils/compare/@kcinternal/react-testkit@0.2.2-rc.0...@kcinternal/react-testkit@0.2.3-rc.1) (2020-06-12)
+
+**Note:** Version bump only for package @kcinternal/react-testkit
+
+
+
+
+
+## [0.2.2-rc.0](https://github.com/kamontat/kcutils/compare/@kcinternal/react-testkit@0.2.2-beta.2...@kcinternal/react-testkit@0.2.2-rc.0) (2020-06-11)
+
+**Note:** Version bump only for package @kcinternal/react-testkit
+
+
+
+
+
+## [0.2.2-beta.2](https://github.com/kamontat/kcutils/compare/@kcinternal/react-testkit@0.2.2-beta.1...@kcinternal/react-testkit@0.2.2-beta.2) (2020-06-08)
+
+**Note:** Version bump only for package @kcinternal/react-testkit
+
+
+
+
+
+## [0.2.2-beta.1](https://github.com/kamontat/kcutils/compare/@kcinternal/react-testkit@0.2.2-beta.0...@kcinternal/react-testkit@0.2.2-beta.1) (2020-06-08)
+
+**Note:** Version bump only for package @kcinternal/react-testkit
+
+
+
+
+
+## 0.2.2-beta.0 (2020-06-08)
+
+
+### Performance Improvements
+
+* **core:** implement changes internal to kcinternal scope ([590d10f](https://github.com/kamontat/kcutils/commit/590d10ff35d617e9964691b7a12d10f5b9170902))
+
+
+
+
+
 ## [0.2.1](https://github.com/kamontat/kcutils/compare/@internal/testkit@0.2.0...@internal/testkit@0.2.1) (2020-06-07)
 
 **Note:** Version bump only for package @internal/testkit