+++ packages/helper/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0-rc.1](https://github.com/kamontat/kcutils/compare/@kcutils/helper@0.2.2-rc.0...@kcutils/helper@0.3.0-rc.1) (2020-06-12)
+
+
+### Bug Fixes
+
+* **lint:** fix code review by sonar ([84a3a07](https://github.com/kamontat/kcutils/commit/84a3a0743a78b66be31db67a7ba8249a92fd720d))
+* **model:** after should return undefined with it not found ([4846859](https://github.com/kamontat/kcutils/commit/4846859a0b4a2a6c3fa9c61ec4b3d674ec64d328))
+* **test:** update coverage on helper to 100% ([d00bd5a](https://github.com/kamontat/kcutils/commit/d00bd5a2a2456cc5a4a1276c63fea64314ea702c))
+* **test:** update test fails ([06970d2](https://github.com/kamontat/kcutils/commit/06970d2f05e4e5591d8a854fad26a3b654c293ee))
+
+
+### Performance Improvements
+
+* **model:** improve new models and how to called ([ce3360e](https://github.com/kamontat/kcutils/commit/ce3360ee5f43f12de779f9a1e9c03e8ac93dd051))
+* **model:** update new stack output for custom throwable class ([157c2a9](https://github.com/kamontat/kcutils/commit/157c2a982a9ae27e0fb44c3c9c75096ddf33b1a4))
+
+
+
+
+
+## [0.2.2-rc.0](https://github.com/kamontat/kcutils/compare/@kcutils/helper@0.2.2-beta.2...@kcutils/helper@0.2.2-rc.0) (2020-06-11)
+
+**Note:** Version bump only for package @kcutils/helper
+
+
+
+
+
+## [0.2.2-beta.2](https://github.com/kamontat/kcutils/compare/@kcutils/helper@0.2.2-beta.1...@kcutils/helper@0.2.2-beta.2) (2020-06-08)
+
+**Note:** Version bump only for package @kcutils/helper
+
+
+
+
+
+## [0.2.2-beta.1](https://github.com/kamontat/kcutils/compare/@kcutils/helper@0.2.2-beta.0...@kcutils/helper@0.2.2-beta.1) (2020-06-08)
+
+**Note:** Version bump only for package @kcutils/helper
+
+
+
+
+
+## [0.2.2-beta.0](https://github.com/kamontat/kcutils/compare/@kcutils/helper@0.2.1...@kcutils/helper@0.2.2-beta.0) (2020-06-08)
+
+
+### Bug Fixes
+
+* **config:** publish denie ([5b9b2de](https://github.com/kamontat/kcutils/commit/5b9b2de3ad51695cda80fc488be3702b08b0afb6))
+
+
+### Performance Improvements
+
+* **core:** implement changes internal to kcinternal scope ([590d10f](https://github.com/kamontat/kcutils/commit/590d10ff35d617e9964691b7a12d10f5b9170902))
+
+
+
+
+
 ## [0.2.1](https://github.com/kamontat/kcutils/compare/@kcutils/helper@0.2.0...@kcutils/helper@0.2.1) (2020-06-07)
 
 **Note:** Version bump only for package @kcutils/helper